azure.dbformysql: add azurerm_mysql_flexible_server
Caveat: didn't manage to make it working with private_dns_zone_id
reference

The message was cryptic:
```
The Private DNS Zone name provided is invalid. It must end with '{0}', and shouldn't contain underscore. Currently we do not support {1}.{0} as the private dns zone name either.
```

Several attempts to guess it and follow official docs failed

Signed-off-by: Yury Tsarev <yury@upbound.io>